refactor: 481476 - Cleanup jsdoc imports move lists out of definition
