+++ src/api/forms/service/definition.js
@@ -305,58 +305,6 @@ export async function reorderDraftFormDefinitionPages(
 }
 
 /**
- * Add a list of new lists to the draft form definition
- * @param {string} formId
- * @param {List[]} lists
- * @param {FormMetadataAuthor} author
- */
-export async function addListsToDraftFormDefinition(formId, lists, author) {
-  logger.info(
-    `Adding lists ${lists.map((list) => list.name).join(', ')} on Form Definition (draft) for form ID ${formId}`
-  )
-
-  const session = client.startSession()
-
-  try {
-    const newForm = await session.withTransaction(async () => {
-      // Add the lists to the form definition
-      const returnedLists = await formDefinition.addLists(
-        formId,
-        lists,
-        session
-      )
-
-      const now = new Date()
-      await formMetadata.update(
-        formId,
-        {
-          $set: partialAuditFields(now, author)
-        },
-        session
-      )
-
-      return returnedLists
-    })
-
-    logger.info(
-      `Added lists ${lists.map((list) => list.name).join(', ')} on Form Definition (draft) for form ID ${formId}`
-    )
-
-    return newForm
-  } catch (err) {
-    logger.error(
-      err,
-      `Failed to add lists ${lists.map((list) => list.id).join(', ')} on Form Definition (draft) for form ID ${formId}`
-    )
-
-    throw err
-  } finally {
-    await session.endSession()
-  }
-}
-
-/**
- * @import { FormDefinition, FormMetadataAuthor, FormMetadata, FilterOptions, QueryOptions, List } from '@defra/forms-model'
+ * @import { FormDefinition, FormMetadataAuthor, FormMetadata, FilterOptions, QueryOptions } from '@defra/forms-model'
  * @import { ClientSession } from 'mongodb'
- * @import { State } from '~/src/api/forms/service/shared.js'
  */
