--- src/api/forms/service/list.js
@@ -1,5 +1,0 @@
-/**
- * @import { FormDefinition, FormMetadataAuthor, FormMetadata, FilterOptions, QueryOptions, List } from '@defra/forms-model'
- * @import { ClientSession } from 'mongodb'
- * @import { State } from '~/src/api/forms/service/shared.js'
- */
